+++ src/Models/Environment.php
@@ -92,7 +92,8 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
             $password = $cache_binding->get('password');
             $domain = $cache_binding->get('host');
             $port = $cache_binding->get('port');
-            $url = "redis://pantheon:$password@$domain:$port";
+            $username = $cache_binding->getUsername();
+            $url = "redis://$username:$password@$domain:$port";
             $command = "redis-cli -h $domain -p $port -a $password";
             return [
                 'password' => $password,
@@ -183,6 +184,7 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
      */
     public function commitChanges($commit = null)
     {
+        $nickname = \uniqid(__FUNCTION__ . "-");
         $local = $this->getContainer()->get(LocalMachineHelper::class);
 
         $git_email_result = $local->exec('git config user.email');
@@ -288,10 +290,10 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
                 $db_binding = array_shift($dbserver_binding);
             } while ($db_binding->get('environment') != $this->id);
 
-            $username = 'pantheon';
             $password = $db_binding->get('password');
             $domain = "dbserver.{$this->id}.{$this->getSite()->id}.drush.in";
             $port = $db_binding->get('port');
+            $username = $db_binding->getUsername();
             $database = 'pantheon';
             $url = "mysql://$username:$password@$domain:$port/$database";
             $command = "mysql -u $username -p$password -h $domain -P $port $database";
@@ -387,10 +389,14 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
     /**
      * @return Backups
      */
-    public function getBackups()
+    public function getBackups() : Backups
     {
         if (empty($this->backups)) {
-            $this->backups = $this->getContainer()->get(Backups::class, [['environment' => $this,],]);
+            $nickname = \uniqid(__FUNCTION__ . "-");
+
+            $this->getContainer()->add($nickname, Backups::class)
+                ->addArguments([['environment' => $this]]);
+            $this->backups = $this->getContainer()->get($nickname);
         }
         return $this->backups;
     }
@@ -401,7 +407,11 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
     public function getBindings()
     {
         if (empty($this->bindings)) {
-            $this->bindings = $this->getContainer()->get(Bindings::class, [['environment' => $this,],]);
+            $nickname = \uniqid(__FUNCTION__ . "-");
+
+            $this->getContainer()->add($nickname, Bindings::class)
+                ->addArguments([['environment' => $this]]);
+            $this->bindings = $this->getContainer()->get($nickname);
         }
         return $this->bindings;
     }
@@ -420,7 +430,11 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
     public function getCommits()
     {
         if (empty($this->commits)) {
-            $this->commits = $this->getContainer()->get(Commits::class, [['environment' => $this,],]);
+            $nickname = \uniqid(__FUNCTION__ . "-");
+
+            $this->getContainer()->add($nickname, Commits::class)
+                ->addArguments([['environment' => $this]]);
+            $this->commits = $this->getContainer()->get($nickname);
         }
         return $this->commits;
     }
@@ -431,9 +445,12 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
     public function getEnvironmentMetrics()
     {
         if (empty($this->environment_metrics)) {
+            $nickname = \uniqid(__FUNCTION__ . "-");
+
+            $this->getContainer()->add($nickname, EnvironmentMetrics::class)
+                ->addArguments([['environment' => $this,],]);
             $this->environment_metrics = $this->getContainer()->get(
-                EnvironmentMetrics::class,
-                [['environment' => $this,],]
+                $nickname
             );
         }
         return $this->environment_metrics;
@@ -445,7 +462,11 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
     public function getDomains()
     {
         if (empty($this->domains)) {
-            $this->domains = $this->getContainer()->get(Domains::class, [['environment' => $this,],]);
+            $nickname = \uniqid(__FUNCTION__ . "-");
+
+            $this->getContainer()->add($nickname, Domains::class)
+                ->addArguments([['environment' => $this]]);
+            $this->domains = $this->getContainer()->get($nickname);
         }
         return $this->domains;
     }
@@ -455,7 +476,11 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
      */
     public function getPrimaryDomainModel()
     {
-        return $this->getContainer()->get(PrimaryDomain::class, [$this]);
+        $nickname = \uniqid(__FUNCTION__ . "-");
+
+        $this->getContainer()->add($nickname, PrimaryDomain::class)
+            ->addArguments([$this]);
+        return $this->getContainer()->get($nickname);
     }
 
     /**
@@ -476,7 +501,10 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
     public function getLock()
     {
         if (empty($this->lock)) {
-            $this->lock = $this->getContainer()->get(Lock::class, [$this->get('lock'), ['environment' => $this],]);
+            $nickname = \uniqid(__FUNCTION__ . "-");
+            $this->getContainer()->add($nickname, Lock::class)
+                ->addArguments([$this->get('lock'), ['environment' => $this]]);
+            $this->lock = $this->getContainer()->get($nickname);
         }
         return $this->lock;
     }
@@ -530,10 +558,10 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
     public function getUpstreamStatus()
     {
         if (empty($this->upstream_status)) {
-            $this->upstream_status = $this->getContainer()->get(
-                UpstreamStatus::class,
-                [[], ['environment' => $this,],]
-            );
+            $nickname = \uniqid(__FUNCTION__ . "-");
+            $this->getContainer()->add($nickname, UpstreamStatus::class)
+                ->addArguments([[], ['environment' => $this,],]);
+            $this->upstream_status = $this->getContainer()->get($nickname);
         }
         return $this->upstream_status;
     }
@@ -544,7 +572,10 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
     public function getWorkflows()
     {
         if (empty($this->workflows)) {
-            $this->workflows = $this->getContainer()->get(Workflows::class, [['environment' => $this,],]);
+            $nickname = \uniqid(__FUNCTION__ . "-");
+            $this->getContainer()->add($nickname, Workflows::class)
+                ->addArguments([['environment' => $this]]);
+            $this->workflows = $this->getContainer()->get($nickname);
         }
         return $this->workflows;
     }
@@ -788,9 +819,11 @@ class Environment extends TerminusModel implements ContainerAwareInterface, Site
             "sites/{$this->getSite()->id}/environments/{$this->id}/add-ssl-cert",
             ['method' => 'POST', 'form_params' => array_filter($certificate),]
         );
-
+        $nickname = \uniqid(__FUNCTION__ . "-");
         // The response is actually a workflow
-        return $this->getContainer()->get(Workflow::class, [$response['data'], ['environment' => $this,],]);
+        $this->getContainer()->add($nickname, Workflow::class)
+            ->addArguments([$response['data'], ['environment' => $this]]);
+        return $this->getContainer()->get($nickname);
     }
 
     /**