+++ src/Models/TerminusModel.php
@@ -69,7 +69,7 @@ abstract class TerminusModel implements ConfigAwareInterface, RequestAwareInterf
      */
     public function fetch(array $args = [])
     {
-        $options = array_merge(['options' => ['method' => 'get',],], $args);
+        $options = array_merge(['options' => ['method' => 'get']], $args);
         $results = $this->request->request($this->getUrl(), $options);
         $this->attributes = (object)array_merge(
             (array)$this->attributes,